refactor: move draft_attribute_renderer
