+++ packages/renderers/lib/src/request/request_item_renderer/decidable/decidable_propose_attribute_request_item_renderer.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 
 import '../../request_item_index.dart';
 import '../../request_renderer_controller.dart';
-import '../widgets/draft_attribute_renderer.dart';
 import '/src/attribute/attribute_renderer.dart';
+import '/src/attribute/draft_attribute_renderer.dart';
 import 'checkbox_enabled_extension.dart';
 import 'widgets/handle_checkbox_change.dart';
